+++ CHANGELOG.md
@@ -428,4 +428,7 @@ Version 0.2.30 (TBD)
 * Add explicit mappings for <=, >=, >, !=
 * Make compound scalar objects match UnNamedCompoundScalar objects with the same properties.
 * add extraction of a single ts value from a data-frame source
-* to_frame operator with to_frame_ts
+* to_frame operator with:
+  * to_frame_ts
+  * to_frame_tsb
+